mpi_examples/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile mpi_examples/requirements.in
#
attrs==21.4.0
    # via jsonschema
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
idna==3.3
    # via requests
jsonschema==4.6.0
    # via valohai-yaml
pycryptodome==3.14.1
    # via -r mpi_examples/requirements.in
pyrsistent==0.18.1
    # via jsonschema
pyyaml==6.0
    # via valohai-yaml
requests==2.27.1
    # via valohai-utils
tqdm==4.64.0
    # via valohai-utils
urllib3==1.26.9
    # via requests
valohai-papi==0.1.3
    # via valohai-utils
valohai-utils==0.2.0
    # via -r mpi_examples/requirements.in
valohai-yaml==0.23.0
    # via
    #   valohai-papi
    #   valohai-utils
